refactor: move safe_partition_analysis to result utils
